This commit will resolve problem with empty USER env variable on CI
In program we use getenv(USER) to get local username. In CI this env variable is empty and all commands are runs by root, so we have different behavior locally and on CI and this causes test failures. 'export USER=ciGitlab' in CI script and getUser() function in UT resolve this problem and now UT pass localy and on CI